Moved add transaction to own component Allow adding transaction from transactions view
