+++ src/app/transactions-view/transactions-view.component.html
@@ -1,3 +1,4 @@
 <h1>{{month}}</h1>
+<div><button mat-button (click)="addTransaction()">Add Transaction</button></div>
 <app-transactions-table [dataSource]="transactions"></app-transactions-table>
